Addition of most JavaDocs and removal of deprecated/unused code
