+++ src/main/java/seedu/planner/logic/commands/student/StudentActiveCommand.java
@@ -40,6 +40,12 @@ public class StudentActiveCommand extends StudentCommand {
         return String.format(MESSAGE_ACTIVE_STUDENT_SUCCESS, student);
     }
 
+    /**
+     * Executes {@code StudentActiveCommand}
+     * @param model {@code Model} which the command should operate on
+     * @return {@code CommandResult} produced from executing the command
+     * @throws CommandException when an exception occurred while executing the command
+     */
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
