+++ src/main/java/seedu/planner/logic/commands/student/StudentListCommand.java
@@ -42,6 +42,12 @@ public class StudentListCommand extends StudentCommand {
         return String.format(MESSAGE_SUCCESS, sb.length() == 0 ? "[None]" : sb.toString());
     }
 
+    /**
+     * Executes {@code StudentListCommand}
+     * @param model {@code Model} which the command should operate on
+     * @return {@code CommandResult} produced from executing the command
+     * @throws CommandException when an exception occurred while executing the command
+     */
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
