+++ src/main/java/seedu/planner/model/graduation/CompoundGraduationRequirement.java
@@ -10,13 +10,36 @@ import seedu.planner.model.module.Module;
 import seedu.planner.model.module.ModuleCode;
 import seedu.planner.model.util.ModuleUtil;
 
+/**
+ * Class that implements {@code GraduationRequirement}, consisting of multiple other {@code GraduationRequirement}
+ * classes. Has an {@code AggregationType} to determine if it has fulfilled requirements
+ */
 public class CompoundGraduationRequirement extends GraduationRequirement {
 
+    /**
+     * Enum Aggregation Type
+     */
     protected AggregationType aggregationType;
+    /**
+     * Name of {@code CompoundAggregationType}
+     */
     protected String name;
+    /**
+     * Minimum number of ModuleCredits, applicable only when {@code AggregationType} is AT_LEAST_MC
+     */
     protected int minMCs;
+    /**
+     * List of {@code GraduationRequirement}
+     */
     protected List<GraduationRequirement> graduationRequirementList = new ArrayList<>();
 
+    /**
+     * Default constructor for {@code CompoundGraduationRequirement}
+     *
+     * @param name String label for {@code CompoundGraduationRequirement}
+     * @param minMCs Minimum number of ModuleCredits, applicable only when {@code AggregationType} is AT_LEAST_MC
+     * @param requirements List of {@code GraduationRequirement}
+     */
     public CompoundGraduationRequirement(String name, int minMCs, List<GraduationRequirement> requirements) {
         this.name = name;
         this.minMCs = minMCs;
@@ -24,6 +47,14 @@ public class CompoundGraduationRequirement extends GraduationRequirement {
         aggregationType = ALL;
     }
 
+    /**
+     * Constructor for {@code CompoundGraduationRequirement}
+     *
+     * @param name String label for {@code CompoundGraduationRequirement}
+     * @param minMCs Minimum number of ModuleCredits, applicable only when {@code AggregationType} is AT_LEAST_MC
+     * @param requirements List of {@code GraduationRequirement}
+     * @param aggregationType Aggregation Type to determine if the @code CompoundGraduationRequirement} is fulfilled
+     */
     public CompoundGraduationRequirement(String name, int minMCs, List<GraduationRequirement> requirements,
                                          AggregationType aggregationType) {
         this.name = name;
@@ -32,10 +63,22 @@ public class CompoundGraduationRequirement extends GraduationRequirement {
         this.aggregationType = aggregationType;
     }
 
+    /**
+     * Returns list of {@code GradudationRequirement}
+     *
+     * @return List of {@code GradudationRequirement}
+     */
     public List<GraduationRequirement> getGraduationRequirementList() {
         return graduationRequirementList;
     }
 
+    /**
+     * Returns a boolean representing if the {@code CompoundGraduationRequirement} is fulfilled, given a list of
+     * {@code ModuleCode}
+     *
+     * @param moduleCodes List of {@code ModuleCode}
+     * @return True if fulfilled. False otherwise.
+     */
     public boolean isFulfilled(List<ModuleCode> moduleCodes) {
         if (graduationRequirementList == null) {
             return false;
@@ -59,9 +102,14 @@ public class CompoundGraduationRequirement extends GraduationRequirement {
             int currentMc = 0;
             for (GraduationRequirement requirement : graduationRequirementList) {
                 if (requirement.isFulfilled(moduleCodes)) {
-                    Module module = ModuleUtil.getModuleWithCode(requirement.getModuleCode());
-                    if (module != null) {
-                        currentMc += module.getModuleCredit();
+                    try {
+                        Module module = ModuleUtil.getModuleWithCode(((SingleGraduationRequirement) requirement)
+                                .getModuleCode());
+                        if (module != null) {
+                            currentMc += module.getModuleCredit();
+                        }
+                    } catch (Exception e) {
+                        continue;
                     }
                 }
             }
@@ -71,6 +119,13 @@ public class CompoundGraduationRequirement extends GraduationRequirement {
         }
     }
 
+    /**
+     * Returns a String representation of the {@code CompoundGraduationRequirement} object, along with it's child
+     * {@code GraduationRequirement}s
+     *
+     * @param moduleCodes List of {@code ModuleCode}
+     * @return A String representation of the {@code CompoundGraduationRequirement} object, along with it's child
+     */
     public String getString(List<ModuleCode> moduleCodes) {
         StringBuilder sb = new StringBuilder();
         String buffer;
@@ -93,6 +148,11 @@ public class CompoundGraduationRequirement extends GraduationRequirement {
         return sb.toString();
     }
 
+    /**
+     * Method to override the default {@code toString} function of {@code FocusAreaGraduationRequirement}
+     *
+     * @return String representation of {@code FocusAreaGraduationRequirement}
+     */
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
