+++ src/main/java/seedu/planner/model/graduation/GraduationRequirement.java
@@ -4,30 +4,42 @@ import java.util.List;
 
 import seedu.planner.model.module.ModuleCode;
 
+/**
+ * Abstract class used for the processing of requirements needed to graduation in a given {@DegreeProgramme}
+ */
 public abstract class GraduationRequirement {
 
-    protected ModuleCode moduleCode;
 
-    public ModuleCode getModuleCode() {
-        return moduleCode;
-    }
-
-    public void setModuleCode(ModuleCode moduleCode) {
-        this.moduleCode = moduleCode;
-    }
-
+    /**
+     * Returns a boolean representing if the {@code GraduationRequirement} is fulfilled, given a list of
+     * {@code ModuleCode}
+     * @param moduleCodes List of {@code ModuleCode}
+     * @return True if fulfilled. False otherwise.
+     */
     public abstract boolean isFulfilled(List<ModuleCode> moduleCodes);
 
+    /**
+     * Helper function to return a tick or cross icon to represent if a graduation has been fulfilled
+     * @param done Boolean of whether the requirement is fulfilled or not
+     * @return A String representation of a unicode tick or cross symbol
+     */
     protected String getStatusIcon(boolean done) {
         return done ? "\u2713" : "\u2718";
     }
 
+    /**
+     * Abstract method to override the default {@code toString} function of {@code GraduationRequirement}
+     *
+     * @return String representation of {@code GraduationRequirement}
+     */
     @Override
-    public String toString() {
-        return moduleCode.toString();
-    }
+    public abstract String toString();
 
-    public String getString(List<ModuleCode> moduleCodes) {
-        return "[" + getStatusIcon(isFulfilled(moduleCodes)) + "] " + moduleCode.toString();
-    }
+    /**
+     * Abstract method to return a String representation of the {@code GraduationRequirement} object
+     *
+     * @param moduleCodes List of {@code ModuleCode}
+     * @return A String representation of the {@code GraduationRequirement} object
+     */
+    public abstract String getString(List<ModuleCode> moduleCodes);
 }
